pyproject.toml
[project]
name = "aaronson-oracle-baseball"
version = "0.1.0"
description = "Adapting Aaronson's Oracle to predict baseball pitches using n-gram pattern matching"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "numpy>=2.1.1",
    "pandas>=2.2.3",
    "pybaseball>=2.2.7",
    "redis>=5.0.8",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "MLB-StatsAPI>=1.7.0",
]

[project.optional-dependencies]
training = [
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "torch>=2.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*", "training*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
